Separate CSS timing and allocation benchmarks
